+++ packages/firebase_auth/firebase_auth_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0
+
+- Added type `PlatformOAuthCredential` for generic OAuth providers.
+
 ## 1.0.0
 
 - Initial open-source release.