+++ packages/firebase_messaging/CHANGELOG.md
@@ -1,6 +1,10 @@
-## 6.0.4
+## 6.0.5
 
 * Add import for UserNotifications on iOS.
+
+## 6.0.4
+
+* Support the v2 Android embedding.
 
 ## 6.0.3